requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
aiohttp>=3.9.0
python-multipart>=0.0.6
PyYAML>=6.0.0
brotli>=1.0.0
